aws-agentcore/pyproject.toml
[tool.poetry]
name = "aws-agentcore"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
common-lib = { path = "../common-lib", develop=true }
python = "^3.11,<3.13"
bedrock-agentcore-starter-toolkit = "^0.1.0"
langgraph = "^0.5.3"
langchain = "^0.3.26"
langchain-anthropic = "^0.3.17"
